Move setLexers to base
